requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
sqlalchemy==2.0.35
asyncpg==0.29.0
pydantic==2.9.0
python-telegram-bot==21.0
celery==5.4.0
redis==5.0.0
python-dotenv==1.0.0
alembic==1.13.0
pgvector==0.2.5
prometheus-fastapi-instrumentator==7.0.0
PyGithub==2.3.0
psycopg2-binary==2.9.12
matplotlib==3.9.0
pandas==2.2.0
# Дополнительно для Фазы 5
cryptography==42.0.0          # для шифрования данных
reportlab==4.2.0              # для генерации PDF-договоров и чеков
prometheus-client==0.19.0     # метрики
python-multipart==0.0.9       # для загрузки файлов